+++ NEWS.md
@@ -1,5 +1,8 @@
 # CHANGES IN pagedown VERSION 0.7
 
+## NEW FEATURES
+
+- Added support for pagebreaks: in output formats which use **paged.js**, a pagebreak can be forced using the LaTeX commands `\newpage` and `\pagebreak` or using the CSS classes `page-break-before` and `page-break-after`.
 
 # CHANGES IN pagedown VERSION 0.6