+++ NEWS.md
@@ -1,3 +1,7 @@
+# CHANGES IN pagedown VERSION 0.8
+
+
+
 # CHANGES IN pagedown VERSION 0.7
 
 ## NEW FEATURES